changed db models
